+++ ChangeLog.md
@@ -25,6 +25,7 @@
     - Homogeneize the 'NOTOK' vs 'NOT OK' tests results
     - Fix the count of elementary tests in t-pl-commands
     - DBMS.limitDatabases and DBMS.excludeDatabases accept a Perl regular expression (todo #152)
+    - Define new DBMS.matchInsensitive property to match database names (todo #153)
 
 ### 4.13.0